add stats and login ready

+++ src/main/java/com/topicos/backend/controller/AreaController.java
@@ -2,6 +2,8 @@ package com.topicos.backend.controller;
 
 import com.topicos.backend.dto.AreaDTO;
 import com.topicos.backend.dto.request.AreaRequestDTO;
+import com.topicos.backend.exceptions.UnauthorizedException;
+import com.topicos.backend.security.JwtTokenUtil;
 import com.topicos.backend.services.AreaService;
 import java.util.List;
 import lombok.AllArgsConstructor;
@@ -23,28 +25,39 @@ public class AreaController {
 
   private final AreaService areaService;
 
+  private final JwtTokenUtil jwtTokenUtil;
+
   //CREATE
   @PostMapping("/area/create")
   public AreaDTO addArea(@RequestBody AreaRequestDTO areaDTO, @RequestHeader("Authorization") String token) {
-    return this.areaService.addArea(areaDTO);
+    if (this.jwtTokenUtil.getAdminFromToken(token)) {
+      return this.areaService.addArea(areaDTO);
+    }
+    throw new UnauthorizedException("The user is not an admin", "The user is not an admin");
   }
 
   //DELETE
   @DeleteMapping("/area/delete")
   public void deleteArea(@RequestParam Long id, @RequestHeader("Authorization") String token) {
-    this.areaService.deleteArea(id);
+    if (this.jwtTokenUtil.getAdminFromToken(token)) {
+      this.areaService.deleteArea(id);
+    }
+    throw new UnauthorizedException("The user is not an admin", "The user is not an admin");
   }
 
   //GET
   @GetMapping("/area")
-  public List<AreaDTO> getAllAreas(Long companyId, @RequestHeader("Authorization") String token) {
-    return this.areaService.getAllAreas(companyId);
+  public List<AreaDTO> getAllAreas(@RequestHeader("Authorization") String token) {
+    return this.areaService.getAllAreas();
   }
 
-  //MODIFICATE
+  //MODIFICATION
   @PutMapping("/area/modify")
   public AreaDTO modifyArea(@RequestBody AreaDTO areaDTO, @RequestHeader("Authorization") String token) {
-    return this.areaService.modifyArea(areaDTO);
+    if (this.jwtTokenUtil.getAdminFromToken(token)) {
+      return this.areaService.modifyArea(areaDTO);
+    }
+    throw new UnauthorizedException("The user is not an admin", "The user is not an admin");
   }
 
 }
