+++ src/main/java/com/topicos/backend/controller/IndicatorController.java
@@ -2,6 +2,8 @@ package com.topicos.backend.controller;
 
 import com.topicos.backend.dto.IndicatorDTO;
 import com.topicos.backend.dto.request.IndicatorRequestDTO;
+import com.topicos.backend.exceptions.UnauthorizedException;
+import com.topicos.backend.security.JwtTokenUtil;
 import com.topicos.backend.services.IndicatorService;
 import java.util.List;
 import lombok.AllArgsConstructor;
@@ -23,16 +25,26 @@ public class IndicatorController {
 
   private final IndicatorService indicatorService;
 
+  private final JwtTokenUtil jwtTokenUtil;
+
   //CREATE
   @PostMapping("/indicator/create")
   public IndicatorDTO addIndicator(@RequestBody IndicatorRequestDTO indicator, @RequestHeader("Authorization") String token) {
-    return this.indicatorService.createIndicator(indicator);
+    if (this.jwtTokenUtil.getAdminFromToken(token)) {
+      return this.indicatorService.createIndicator(indicator);
+    }
+    throw new UnauthorizedException("The user is not an admin", "The user is not an admin");
+
   }
 
   //DELETE
   @DeleteMapping("/indicator/delete")
   public void deleteIndicator(@RequestParam Long id, @RequestHeader("Authorization") String token) {
-    this.indicatorService.deleteIndicator(id);
+    if (this.jwtTokenUtil.getAdminFromToken(token)) {
+      this.indicatorService.deleteIndicator(id);
+    }
+    throw new UnauthorizedException("The user is not an admin", "The user is not an admin");
+
   }
 
   //GET
@@ -41,10 +53,13 @@ public class IndicatorController {
     return this.indicatorService.getAllIndicators();
   }
 
-  //MODIFICATE
+  //MODIFICATION
   @PutMapping("/indicator/modify")
   public IndicatorDTO modifyIndicator(@RequestBody IndicatorRequestDTO indicator, @RequestHeader("Authorization") String token) {
-    return this.indicatorService.modifyIndicator(indicator);
+    if (this.jwtTokenUtil.getAdminFromToken(token)) {
+      return this.indicatorService.modifyIndicator(indicator);
+    }
+    throw new UnauthorizedException("The user is not an admin", "The user is not an admin");
   }
 
 }
